feat(build)!: add forward declaration header (#203)
* added Forward.hpp


* style(pre-commit): autofix

---------
