+++ lanelet2_extension/include/lanelet2_extension/regulatory_elements/speed_bump.hpp
@@ -19,11 +19,11 @@
 
 // NOLINTBEGIN(readability-identifier-naming)
 
-#include <lanelet2_core/primitives/BasicRegulatoryElements.h>
+#include <lanelet2_extension/regulatory_elements/Forward.hpp>
+
 #include <lanelet2_core/primitives/Lanelet.h>
 
 #include <memory>
-#include <vector>
 
 namespace lanelet::autoware
 {
