+++ lanelet2_extension/include/lanelet2_extension/regulatory_elements/virtual_traffic_light.hpp
@@ -17,11 +17,11 @@
 
 // NOLINTBEGIN(readability-identifier-naming)
 
-#include <lanelet2_core/primitives/BasicRegulatoryElements.h>
+#include <lanelet2_extension/regulatory_elements/Forward.hpp>
+
 #include <lanelet2_core/primitives/LineString.h>
 
 #include <memory>
-#include <vector>
 
 namespace lanelet::autoware
 {
