+++ lanelet2_extension/include/lanelet2_extension/utility/query.hpp
@@ -34,22 +34,8 @@
 #include <lanelet2_routing/RoutingGraph.h>
 
 #include <limits>
-#include <memory>
 #include <string>
 #include <vector>
-
-namespace lanelet
-{
-using TrafficSignConstPtr = std::shared_ptr<const lanelet::TrafficSign>;
-using TrafficLightConstPtr = std::shared_ptr<const lanelet::TrafficLight>;
-using AutowareTrafficLightConstPtr = std::shared_ptr<const lanelet::autoware::AutowareTrafficLight>;
-using DetectionAreaConstPtr = std::shared_ptr<const lanelet::autoware::DetectionArea>;
-using NoParkingAreaConstPtr = std::shared_ptr<const lanelet::autoware::NoParkingArea>;
-using NoStoppingAreaConstPtr = std::shared_ptr<const lanelet::autoware::NoStoppingArea>;
-using NoParkingAreaConstPtr = std::shared_ptr<const lanelet::autoware::NoParkingArea>;
-using SpeedBumpConstPtr = std::shared_ptr<const lanelet::autoware::SpeedBump>;
-using CrosswalkConstPtr = std::shared_ptr<const lanelet::autoware::Crosswalk>;
-}  // namespace lanelet
 
 namespace lanelet::utils::query
 {
