+++ lanelet2_extension/lib/message_conversion.cpp
@@ -23,6 +23,7 @@
 #include <boost/archive/binary_iarchive.hpp>
 #include <boost/archive/binary_oarchive.hpp>
 
+#include <lanelet2_core/LaneletMap.h>
 #include <lanelet2_core/primitives/Lanelet.h>
 #include <lanelet2_io/Exceptions.h>
 #include <lanelet2_io/Projection.h>
@@ -30,6 +31,7 @@
 #include <lanelet2_io/io_handlers/OsmHandler.h>
 #include <lanelet2_io/io_handlers/Serialize.h>
 #include <lanelet2_projection/UTM.h>
+#include <lanelet2_routing/RoutingGraph.h>
 
 #include <sstream>
 #include <string>
