+++ lanelet2_extension/test/src/test_query.cpp
@@ -14,9 +14,11 @@
 
 // NOLINTBEGIN(readability-identifier-naming)
 
+#include "lanelet2_extension/regulatory_elements/autoware_traffic_light.hpp"
 #include "lanelet2_extension/utility/query.hpp"
 
 #include <gtest/gtest.h>
+#include <lanelet2_core/LaneletMap.h>
 
 #include <cmath>
