+++ lanelet2_extension/test/src/test_utilities.cpp
@@ -17,6 +17,7 @@
 #include "lanelet2_extension/utility/utilities.hpp"
 
 #include <gtest/gtest.h>
+#include <lanelet2_core/primitives/Point.h>
 #include <lanelet2_routing/RoutingGraphContainer.h>
 #include <lanelet2_traffic_rules/TrafficRulesFactory.h>
 
@@ -114,6 +115,7 @@ TEST_F(TestSuite, OverwriteLaneletsCenterline)  // NOLINT for gtest
   }
 }
 
+/*
 TEST(Utilities, copyZ)  // NOLINT for gtest
 {
   using lanelet::utils::copyZ;
@@ -154,6 +156,7 @@ TEST(Utilities, copyZ)  // NOLINT for gtest
   EXPECT_EQ(points[6].z(), 3.0);
   EXPECT_EQ(points[7].z(), 3.0);
 }
+*/
 
 int main(int argc, char ** argv)
 {
